pyproject.toml
[project]
name = "wsjf-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.89.0",
    "httpx>=0.28.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "rich>=14.3.3",
    "tabulate>=0.9.0",
]
